Create Makefile to generate elvi from a spec
This is a simple example to generate elvi using a Makefile, a single
text file, and some ubiquitous commands (in the Makefile).

Note: An option to *not* url-escape the arguments may need to be added.
The results for `mal*` (elvi for MyAnimeList) such as for the input
`./malanime tokyo ghoul` are "tokyo%20ghoul", regardless of it is in a
graphical or text browser. Interestingly, my manually written surfraw
for MyAnimeList (myanimelist) does not have this problem.