Add `--member` option
This is the best name I could come up with. "Switch" couldn't be used
because it is synonymous with "flag"... sometimes. At other times
"switch" is given a different meaning to "flag", so I used a different
word to match the meaning of the option.

This has yet to be used in any of my personal elvi but I think I know of
specific elvi that would benefit from this easier use of enum option.