Add list option type
This is a kind of `--enum` option, and therefore had to allow an
optional last arg for `--enum` to be backwards-compatible.  This is very
ugly, but seems to work OK.

The first iteration of mapping being treated specially was removed.  It
would have been too complex with normal *and* list mappings.  Most
websites should be able to handle that (citation needed).

TODO: Refactor to not be so brittle.