Allow specification of option descriptions
This doesn't handle descriptions for the *values* of options.  Right
now, this would only apply to enum options and enum list options.  Other
option types may be added in the future that would also display their
values, for which custom descriptions would be helpful.

TODO: Figure out a general solution for custom descriptions for the
*values* of options.