Add `--inline` and `--list-inline` options
This allows for complex syntax to be uniformly presented to the user as
part of the surfraw elvis interface.  I'm not sure how useful this will
be though.  My attempt at making one for searching github repositories
wasn't fun.

The benefit of this is that the users of elvi would be able to use all
the features of the search engines with the same interface--and not
really being required to read up on the syntax.  Some common conventions
are usually followed:

	1. filetype: `filetype:pdf`
	2. site: `site:example.com`, `site:us` (all .us domains)

But there are some other search syntaxes:

	1. hints: `-dogs` (no dogs), `+dogs` (more dogs) -> both from duckduckgo
	2. bangs: `!wikipedia` or `!w` (search wikipedia) -> duckduckgo

A general solution to this would be nice.