Simply `_ChainContainer` types
Per the Unix guideline of folding knowledge into data, the subclasses of
`_ChainContainer` shouldn't need to have so much code.  Declaring their
contents with some simple data structure in a class attribute should be
enough (like with parsing strings to `Option` instances).

The order of the resolvers for flags and aliases are significant now.
Don't forget!