Split surfraw option classes in two
`Option` subclasses are purely for modelling the input from the user
while `SurfrawOption` subclasses deal with the "business logic", so to
speak.

`self.valid_enum_values` was renamed to `self.values` for list options
since it's more general.  For anything-lists, it's just an empty list;
for enum-lists, it's a list of valid values.  I'd like to extend this in
the future (maybe lists of numeric ranges for a number type?).

`SurfrawAlias` does not have a `type` attribute since it's not needed
anymore--it already has a reference to its target, so knowing its type
should be easy.

The last arg of list options are no longer unlimited--so it has 3 args +
1 optional arg.  This is for easier changes--without breaking stuff--in
the future.

Since I'm using postponed evaluation for annotations from `__future__`
and `dataclasses`, python 3.7 is now the minimum version.  Oh well.