Don't use dataclasses for `SurfrawOption` types
It's a shame that the brevity allowed by dataclasses is gone, but
startup time is important.  It's also pretty cool to see the program run
in under 100 ms (~90 ms for `mkelvis --help`, ~80 ms for calls with many
arguments; how odd).

I have to see what the `__hash__` method of `SurfrawAlias` should be.
It seems to work fine though, given how the program actually operates.