Set development status to alpha
According to [1], I believe that the program more closely matches the
alpha phase.  Near the beginning it was clear what kind of program this
was going to be.  Now that the speed is passable and deprecated options
were removed (replaced with an more complex implementation but better UI
methinks), it feels coherent.

It's not beta because I think there's room for more features and
polishing.  There's also room for other related command-line programs.

[1]: https://martin-thoma.com/software-development-stages/