Apply REUSE guidelines
See the website [1] for more.  I'm not sure whether the licence headers
for `surfraw_tools/templates/elvis.in` should be output to the generated
file.

Should the licence of `MANIFEST.in` be CC0-1.0?  Perhaps it should be
the same as the files included in an sdist?

The `COPYING` file remains in the repo because a repo-wide licence file
seems to be what's read.  A notice in the README should be included too.

The `license_file` key was removed from `setup.cfg`.  I don't know
whether it should be there anymore because of `LICENSES/`.

[1]: https://reuse.software