opensearch2elvis: Support POST queries, custom params, and pre-draft 3
POST method queries just use a GET request and hope for the best.
Doing #27 would make this more robust.

Custom params have two limitations:

1. only expressed as "anything" options; and
2. assume that the XML prefix used to generate the option name is the
same as all the other params that share the same XML namespace.

These should be fixed in the future.

This was a pain to do.  As said in #24, using `Param` elements
*unprefixed* was a feature of the core standard.  This changed from
draft 3, which moved it to the "Parameters" extension and renamed
`Param` to `Parameter`.  A big problem is that the XML namespace for the
core OpenSearch standard remains the same, despite this difference.