lib, mkelvis: Allow disabling appending mappings to search URLs
Fixes #34.

"Allow disabling appending mappings...": now that's a mouthful.  Can't
make it much shorter though.

This also removes restrictions on option combinations because it's
honestly a pain to think about.  The default options are fine as they
are and the advanced user would be fine.  The output is human-readable
too, so it shouldn't be too hard to debug elvi iteratively.